@@ -1,12 +1,13 @@
 ﻿using System.Data.SqlClient;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using Readr.Models;
 using Readr.Repositories;
-using Readr.Repositories.Contexts;
 using Readr.Repositories.Interfaces;
 using Services;
 
@@ -14,11 +15,13 @@ namespace Readr.Api
 {
     public class Startup
     {
+#pragma warning disable CS0618 // IHostingEnvironment is obsolete; use IHostEnvironment instead
         private readonly IHostingEnvironment _env;
         private IConfiguration _config;
         private readonly ILoggerFactory _loggerFactory;
 
         private string _connection = null;
+
 
         public Startup(IHostingEnvironment env, ILoggerFactory loggerFactory)
         {
@@ -38,20 +41,19 @@ namespace Readr.Api
             _env = env;
             _config = builder.Build();
         }
+#pragma warning restore CS0618 // IHostingEnvironment is obsolete; use IHostEnvironment instead
 
         public void ConfigureServices(IServiceCollection services)
         {
-            //services.AddRazorPages();
+            //add MongoDB configurations
+            services.Configure<MongoDbSettings>(
+                _config.GetSection(nameof(MongoDbSettings)));
 
-            var builder = new SqlConnectionStringBuilder(
-            _config["connectionString"]);
-            builder.Password = _config["DbPassword"];
-            _connection = builder.ConnectionString;
+            services.AddSingleton<IMongoDbSettings>(sp => 
+                sp.GetRequiredService<IOptions<MongoDbSettings>>().Value);
 
-            //add DbContext
-            services.AddMvc();
-            services.AddDbContext<AppUserContext>(options =>
-            options.UseSqlServer(_connection));
+            services.AddMvc()
+                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
 
             //handle dependency injections
             services.AddScoped<IAppUserService, AppUserService>();
@@ -60,8 +62,6 @@ namespace Readr.Api
 
         public void Configure(IApplicationBuilder app)
         {
-            //_env = env;
-            //_loggerFactory = loggerFactory;
             var logger = _loggerFactory.CreateLogger<Startup>();
 
             if (_env.IsDevelopment())
@@ -77,26 +77,10 @@ namespace Readr.Api
 
                 // Non-development service configuration
                 logger.LogInformation($"Environment: {_env.EnvironmentName}");
-                
-                //not sure what HSTS is at this time. may be only useful for production environments
-                //app.UseHsts();
             }
-
-            //app.UseHttpsRedirection();
-            //app.UseStaticFiles();
-            //app.UseRouting();
-            //app.UseAuthorization();
-
 
             //Use for MVC 
             app.UseMvc();
-
-
-            //Use for Razor Pages
-            //app.UseEndpoints(endpoints =>
-            //{
-            //    endpoints.MapRazorPages();
-            //});
         }
     }
 }
